♻️ Refactor: 트랜잭셔널 설정 및 기타 클래스들을 규격에 맞게 변경하였습니다.
